requirements.txt
six==1.12.0
requests==2.21.0
pycrypto==2.6.1